feat(view): UserAvatarをRNに対応
